+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## [Unreleased]
 
+## 0.5.0 - (unreleased)
+
+### Added
+
+* Added `Regioncam1D`, `Plane1D`
+
+### Changed
+
+* Generalized `Plane` to `Hyperplane`
+* Renamed `Plane::inverse` to `Hyperplane::project`
+
+
 ## 0.4.1 - 2025-04-17
 
 ### Added
@@ -13,6 +25,7 @@
 ### Changed
 
 * `Regioncam::from_plane` now takes `size` argument
+
 
 ## 0.4.0 - 2025-04-13
 
@@ -27,6 +40,7 @@
 * Face colors are based on hash of activation pattern, this makes colors stable even when other faces change.
 * Remove `_repr_svg_` in python interface in favor of `_repr_png_`, this should result in smaller notebook files.
 
+
 ## 0.3.0 - 2025-02-12
 
 ### Added
@@ -34,12 +48,14 @@
 * Construct a Plane from a Linear transformation
 * Expose sequence of vertices, faces, edges, layers in python interface
 
+
 ## 0.2.0 - 2025-02-10
 
 ### Changed
 
 * Split regioncam into partition (the halfedge data structure) and regioncam (which stores layer outputs)
 * Don't print 0 size points/labels
+
 
 ## 0.1.1 - 2025-01-28